Fixed Lobby connection from multiple Clients
Issue with numPlayer counter in communication class where when a client connects to the lobby, it adds 2 players to the counter rather than one. So the First player joins and it is 1, then 2nd player joins and it's 3, then 3rd player joins and it's 5. Has to do with connection length, not counter itself.